Downloading the json and saving it as shapefiles can now only be done by opening the plugin without 3 layers named manholes, pipes and measuring_points.
