+++ beoordelingstool_dockwidget_base.ui
@@ -36,103 +36,6 @@
      <attribute name="title">
       <string>General</string>
      </attribute>
-     <widget class="QGroupBox" name="groupbox_download">
-      <property name="geometry">
-       <rect>
-        <x>10</x>
-        <y>10</y>
-        <width>411</width>
-        <height>71</height>
-       </rect>
-      </property>
-      <property name="title">
-       <string>Download</string>
-      </property>
-      <widget class="QLabel" name="download_riool_label">
-       <property name="geometry">
-        <rect>
-         <x>0</x>
-         <y>30</y>
-         <width>141</width>
-         <height>21</height>
-        </rect>
-       </property>
-       <property name="text">
-        <string>Putten en leidingen</string>
-       </property>
-      </widget>
-      <widget class="QLineEdit" name="download_riool_text">
-       <property name="geometry">
-        <rect>
-         <x>150</x>
-         <y>30</y>
-         <width>151</width>
-         <height>27</height>
-        </rect>
-       </property>
-       <property name="text">
-        <string>/home/madeleine/code/QGIS_plugins/Beoordelingstool_extra/temp_new.json</string>
-       </property>
-       <property name="readOnly">
-        <bool>true</bool>
-       </property>
-       <property name="placeholderText">
-        <string>riool.json</string>
-       </property>
-      </widget>
-      <widget class="QPushButton" name="download_riool_search">
-       <property name="geometry">
-        <rect>
-         <x>310</x>
-         <y>30</y>
-         <width>99</width>
-         <height>27</height>
-        </rect>
-       </property>
-       <property name="text">
-        <string>Zoek...</string>
-       </property>
-      </widget>
-     </widget>
-     <widget class="QGroupBox" name="groupbox_shapefile">
-      <property name="geometry">
-       <rect>
-        <x>10</x>
-        <y>80</y>
-        <width>411</width>
-        <height>121</height>
-       </rect>
-      </property>
-      <property name="title">
-       <string>Shapefile</string>
-      </property>
-      <widget class="QPushButton" name="save_shapefile_putten_button">
-       <property name="geometry">
-        <rect>
-         <x>150</x>
-         <y>30</y>
-         <width>111</width>
-         <height>27</height>
-        </rect>
-       </property>
-       <property name="text">
-        <string>Save putten</string>
-       </property>
-      </widget>
-      <widget class="QPushButton" name="save_shapefile_leidingen_button">
-       <property name="geometry">
-        <rect>
-         <x>150</x>
-         <y>80</y>
-         <width>111</width>
-         <height>27</height>
-        </rect>
-       </property>
-       <property name="text">
-        <string>Save leidingen</string>
-       </property>
-      </widget>
-     </widget>
     </widget>
     <widget class="QWidget" name="tab_manholes">
      <attribute name="title">
